simple-oauth2: swapped version folders
