+++ types/simple-oauth2/simple-oauth2-tests.ts
@@ -18,7 +18,7 @@ const credentials: oauth2lib.ModuleOptions = {
 const oauth2 = oauth2lib.create(credentials);
 
 // #Authorization Code flow
-(() => {
+(async () => {
     // Authorization oauth2 URI
     const authorizationUri = oauth2.authorizationCode.authorizeURL({
         redirect_uri: 'http://localhost:3000/callback',
@@ -36,86 +36,47 @@ const oauth2 = oauth2lib.create(credentials);
         scope: ['<scope1>', '<scope2>']
     };
 
-    // Callbacks
     // Save the access token
-    oauth2.authorizationCode.getToken(tokenConfig, (error, result) => {
-        if (error) {
-            console.log('Access Token Error', error.message);
-            return;
-        }
-
-        const token = oauth2.accessToken.create(result);
-    });
-
-    // Promises
-    // Save the access token
-    oauth2.authorizationCode.getToken(tokenConfig)
-        .then((result) => {
-            const token = oauth2.accessToken.create(result);
-        })
-        .catch((error) => {
-            console.log('Access Token Error', error.message);
-        });
+    try {
+        const result = await oauth2.authorizationCode.getToken(tokenConfig);
+        const accessToken = oauth2.accessToken.create(result);
+    } catch (error) {
+        console.log('Access Token Error', error.message);
+    }
 })();
 
 // #Password Credentials Flow
-(() => {
+(async () => {
     const tokenConfig = {
         username: 'username',
         password: 'password',
         scope: [ '<scope1>', '<scope2>' ],
     };
 
-    // Callbacks
     // Save the access token
-    oauth2.ownerPassword.getToken(tokenConfig, (error, result) => {
-        if (error) {
-            console.log('Access Token Error', error.message);
-            return;
-        }
-
-        const token = oauth2.accessToken.create(result);
-    });
-
-    // Promises
-    // Save the access token
-    oauth2.ownerPassword.getToken(tokenConfig)
-        .then((result) => {
-            const token = oauth2.accessToken.create(result);
-        })
-        .catch((error) => {
-            console.log('Access Token Error', error.message);
-        });
+    try {
+        const result = await oauth2.ownerPassword.getToken(tokenConfig);
+        const accessToken = oauth2.accessToken.create(result);
+    } catch (error) {
+        console.log('Access Token Error', error.message);
+    }
 })();
 
 // #Client Credentials Flow
-(() => {
+(async () => {
     const tokenConfig = {};
 
-    // Callbacks
     // Get the access token object for the client
-    oauth2.clientCredentials.getToken(tokenConfig, (error, result) => {
-        if (error) {
-            console.log('Access Token Error', error.message);
-            return;
-        }
-
-        const token = oauth2.accessToken.create(result);
-    });
-
-    // Promises
-    // Get the access token object for the client
-    oauth2.clientCredentials.getToken(tokenConfig)
-        .then((result) => {
-            const token = oauth2.accessToken.create(result);
-        })
-        .catch((error) => {
-            console.log('Access Token error', error.message);
-        });
+    try {
+        const result = await oauth2.clientCredentials.getToken(tokenConfig);
+        const accessToken = oauth2.accessToken.create(result);
+    } catch (error) {
+        console.log('Access Token error', error.message);
+    }
 })();
 
 // #Access Token object
-(() => {
+(async () => {
     // Sample of a JSON access token (you got it through previous steps)
     const tokenObject = {
         access_token: '<access-token>',
@@ -128,58 +89,42 @@ const oauth2 = oauth2lib.create(credentials);
 
     // Check if the token is expired. If expired it is refreshed.
     if (accessToken.expired()) {
-        // Callbacks
-        accessToken.refresh((error, result) => {
-            accessToken = result;
-        });
-
-        // Promises
-        accessToken.refresh()
-            .then((result) => {
-                accessToken = result;
-            });
+        try {
+            accessToken = await accessToken.refresh();
+        } catch (error) {
+            console.log('Error refreshing access token: ', error.message);
+        }
     }
 
-    // Callbacks
-    // Revoke only the access token
-    accessToken.revoke('access_token', (error) => {
+    // Revoke both access and refresh tokens
+    try {
+        // Revoke only the access token
+        await accessToken.revoke('access_token');
+
         // Session ended. But the refresh_token is still valid.
+        // Revoke the refresh token
+        await accessToken.revoke('refresh_token');
 
-        // Revoke the refresh_token
-        accessToken.revoke('refresh_token', (error) => {
-            console.log('token revoked.');
-        });
-    });
+        console.log('Token revoked');
+    } catch (error) {
+        console.log('Error revoking token: ', error.message);
+    }
 
-    // Promises
-    // Revoke only the access token
-    accessToken.revoke('access_token')
-        .then(() => {
-            // Revoke the refresh token
-            return accessToken.revoke('refresh_token');
-        })
-        .then(() => {
-            console.log('Token revoked');
-        })
-        .catch((error) => {
-            console.log('Error revoking token.', error.message);
-        });
+    // or...
+
+    try {
+        // Revokes both tokens, refresh token is only revoked if the access_token is properly revoked
+        await accessToken.revokeAll();
+    } catch (error) {
+        console.log('Error revoking token: ', error.message);
+    }
 })();
 
 // #Errors
 // not applicable, as those errors about missing authentication codes are already found by the typescript compiler
 
 // (function () {
-//     // Callbacks
-//     oauth2.authorizationCode.getToken({}, (error, token) => {
-//         if (error) {
-//             return console.log(error.message);
-//         }
-//     });
-
-//     // Promises
-//     oauth2.authorizationCode
-//         .getToken({})
+//     oauth2.authorizationCode.getToken({})
 //         .catch((error) => {
 //             console.log(error.message);
 //         });
