+++ types/simple-oauth2/v1/tslint.json
